requirements.txt
Django>=4.2,<5.0
psycopg2-binary>=2.9,<3.0
djangorestframework==3.15.2
drf-yasg==1.21.7
djangorestframework-simplejwt==5.3.1
celery==5.3.0
redis==5.0.1
django-celery-beat==2.7.0